Implement calculateCellsCount function for ingredient cell distribution and update simulation creation logic; adjust molarFraction type to int

+++ api/schemas.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 
 class Ingredient(BaseModel):
   name: str
-  molarFraction: float
+  molarFraction: int
   color: str
 
 class PairParameter(BaseModel):
